task#287: int-test all C8-Composite health-states
